Specify #include for PluginHeader.h without path
I think this would still work for the current VisualStudio build
and would allow the PluginHeader.h location to be overridden by
placing it in the Source directory or by changing the Include
search path. I did this because this #include is currently the
only thing in the code that points to JuceLibraryCode.